debugged record and moved it to model
